# api_docs.json
{
  "swagger": "2.0",
  "info": {
    "title": "World Cup API",
    "description": "This is an API for the World Cup (2018) and Women's World Cup (2019) that outputs match data as JSON."
  },
...
  "paths": {
...
    "/men/matches/country": {
      "get": {
        "summary": "Get matches for any country, by entering their FIFA Code",
        "parameters": [
          {
            "name": "fifa_code",
            "in": "query",
            "description": "Filter matches by FIFA Code",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Men"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      }
    },
...
    "/women/matches/country": {
      "get": {
        "summary": "Get matches for any country, by entering their FIFA Code",
        "parameters": [
          {
            "name": "fifa_code",
            "in": "query",
            "description": "Filter matches by FIFA Code",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Women"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      }
    },
...
  }
}